Search refactor follow-up: SearchForm for frontend-related query context.
