+++ app/lib/frontend/handlers/custom_api.dart
@@ -15,6 +15,7 @@ import '../../package/name_tracker.dart';
 import '../../package/overrides.dart';
 import '../../scorecard/backend.dart';
 import '../../search/search_client.dart';
+import '../../search/search_form.dart';
 import '../../search/search_service.dart';
 import '../../shared/configuration.dart';
 import '../../shared/exceptions.dart';
@@ -245,13 +246,13 @@ Future<shelf.Response> apiHistoryHandler(shelf.Request request) async {
 
 /// Handles requests for /api/search
 Future<shelf.Response> apiSearchHandler(shelf.Request request) async {
-  final searchQuery = parseFrontendSearchQuery(
+  final searchForm = parseFrontendSearchForm(
     request.requestedUri.queryParameters,
     tagsPredicate: TagsPredicate.regularSearch(),
   );
-  final sr = await searchClient.search(searchQuery);
+  final sr = await searchClient.search(searchForm.toServiceQuery());
   final packages = sr.packages.map((ps) => {'package': ps.package}).toList();
-  final hasNextPage = sr.totalCount > searchQuery.limit + searchQuery.offset;
+  final hasNextPage = sr.totalCount > searchForm.limit + searchForm.offset;
   final result = <String, dynamic>{
     'packages': packages,
     if (sr.message != null) 'message': sr.message,
@@ -259,7 +260,7 @@ Future<shelf.Response> apiSearchHandler(shelf.Request request) async {
   if (hasNextPage) {
     final newParams =
         Map<String, dynamic>.from(request.requestedUri.queryParameters);
-    final nextPageIndex = (searchQuery.offset ~/ searchQuery.limit) + 2;
+    final nextPageIndex = (searchForm.offset ~/ searchForm.limit) + 2;
     newParams['page'] = nextPageIndex.toString();
     final nextPageUrl =
         request.requestedUri.replace(queryParameters: newParams).toString();
