+++ app/lib/frontend/handlers/listing.dart
@@ -9,6 +9,7 @@ import 'package:shelf/shelf.dart' as shelf;
 import '../../package/backend.dart';
 import '../../package/models.dart';
 import '../../package/search_adapter.dart';
+import '../../search/search_form.dart';
 import '../../search/search_service.dart';
 import '../../shared/handlers.dart';
 import '../../shared/tags.dart';
@@ -78,24 +79,23 @@ Future<shelf.Response> _packagesHandlerHtmlCore(
   TagsPredicate tagsPredicate,
   String searchPlaceholder,
 }) async {
-  // TODO: use search memcache for all results here or remove search memcache
-  final searchQuery = parseFrontendSearchQuery(
+  final searchForm = parseFrontendSearchForm(
     request.requestedUri.queryParameters,
     sdk: sdk,
     tagsPredicate: tagsPredicate ?? TagsPredicate.regularSearch(),
   );
   final sw = Stopwatch()..start();
-  final searchResult = await searchAdapter.search(searchQuery);
+  final searchResult = await searchAdapter.search(searchForm);
   final int totalCount = searchResult.totalCount;
 
   final links =
-      PageLinks(searchQuery.offset, totalCount, searchQuery: searchQuery);
+      PageLinks(searchForm.offset, totalCount, searchForm: searchForm);
   final result = htmlResponse(
     renderPkgIndexPage(
       searchResult.packages,
       links,
       sdk: sdk,
-      searchQuery: searchQuery,
+      searchForm: searchForm,
       totalCount: totalCount,
       title: title,
       searchPlaceholder: searchPlaceholder,
