+++ app/lib/frontend/templates/package.dart
@@ -13,7 +13,7 @@ import '../../analyzer/analyzer_client.dart';
 import '../../package/models.dart';
 import '../../package/overrides.dart' show devDependencyPackages;
 import '../../scorecard/models.dart';
-import '../../search/search_service.dart';
+import '../../search/search_form.dart';
 import '../../shared/email.dart' show EmailAddress;
 import '../../shared/tags.dart';
 import '../../shared/urls.dart' as urls;
@@ -437,8 +437,7 @@ String _getAuthorsHtml(List<String> authors) {
     if (author.email != null) {
       final escapedEmail = htmlAttrEscape.convert(author.email);
       final emailSearchUrl = htmlAttrEscape.convert(
-          FrontendSearchQuery.parse(query: 'email:${author.email}')
-              .toSearchLink());
+          SearchForm.parse(query: 'email:${author.email}').toSearchLink());
       final text =
           '<a href="$emailSearchUrl" title="Search packages from $escapedName" rel="nofollow">'
           '$escapedEmail'
