+++ app/lib/search/search_service.dart
@@ -10,7 +10,6 @@ import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 
 import '../shared/tags.dart';
-import '../shared/urls.dart';
 
 part 'search_service.g.dart';
 
@@ -220,241 +219,9 @@ String _stringToNull(String v) => (v == null || v.isEmpty) ? null : v;
 List<String> _listToNull(List<String> list) =>
     (list == null || list.isEmpty) ? null : list;
 
-class FrontendSearchQuery {
-  final String query;
-  final ParsedQuery parsedQuery;
-
-  final TagsPredicate tagsPredicate;
-
-  /// The query will match packages where the owners of the package have
-  /// non-empty intersection with the provided list of owners.
-  ///
-  /// Values of this list can be email addresses (usually a single on) or
-  /// publisher ids (may be multiple).
-  final List<String> uploaderOrPublishers;
-
-  final String publisherId;
-  final SearchOrder order;
-  final int offset;
-  final int limit;
-
-  /// True, if packages with is:discontinued tag should be included.
-  final bool includeDiscontinued;
-
-  /// True, if packages with is:unlisted tag should be included.
-  final bool includeUnlisted;
-
-  /// True, if packages which only support dart 1.x should be included.
-  final bool includeLegacy;
-
-  FrontendSearchQuery._({
-    this.query,
-    TagsPredicate tagsPredicate,
-    List<String> uploaderOrPublishers,
-    String publisherId,
-    this.order,
-    this.offset,
-    this.limit,
-    this.includeDiscontinued,
-    this.includeUnlisted,
-    this.includeLegacy,
-  })  : parsedQuery = ParsedQuery._parse(query),
-        tagsPredicate = tagsPredicate ?? TagsPredicate(),
-        uploaderOrPublishers = _listToNull(uploaderOrPublishers),
-        publisherId = _stringToNull(publisherId);
-
-  factory FrontendSearchQuery.parse({
-    String query,
-    String sdk,
-    List<String> runtimes,
-    List<String> platforms,
-    TagsPredicate tagsPredicate,
-    List<String> uploaderOrPublishers,
-    String publisherId,
-    SearchOrder order,
-    int offset = 0,
-    int limit = 10,
-    bool includeDiscontinued = false,
-    bool includeUnlisted = false,
-    bool includeLegacy = false,
-  }) {
-    final q = _stringToNull(query?.trim());
-    tagsPredicate ??= TagsPredicate();
-    final requiredTags = <String>[];
-    if (SdkTagValue.isNotAny(sdk)) {
-      requiredTags.add('sdk:$sdk');
-    }
-    DartSdkRuntime.decodeQueryValues(runtimes)
-        ?.map((v) => 'runtime:$v')
-        ?.forEach(requiredTags.add);
-    platforms
-        ?.where((v) => v.isNotEmpty)
-        ?.map((v) => 'platform:$v')
-        ?.forEach(requiredTags.add);
-    if (requiredTags.isNotEmpty) {
-      tagsPredicate = tagsPredicate
-          .appendPredicate(TagsPredicate(requiredTags: requiredTags));
-    }
-    return FrontendSearchQuery._(
-      query: q,
-      tagsPredicate: tagsPredicate,
-      uploaderOrPublishers: uploaderOrPublishers,
-      publisherId: publisherId,
-      order: order,
-      offset: offset,
-      limit: limit,
-      includeDiscontinued: includeDiscontinued,
-      includeUnlisted: includeUnlisted,
-      includeLegacy: includeLegacy,
-    );
-  }
-
-  FrontendSearchQuery change({
-    String query,
-    String sdk,
-    TagsPredicate tagsPredicate,
-    List<String> uploaderOrPublishers,
-    String publisherId,
-    SearchOrder order,
-    int offset,
-    int limit,
-  }) {
-    if (sdk != null) {
-      tagsPredicate ??= this.tagsPredicate ?? TagsPredicate();
-      tagsPredicate = tagsPredicate.removePrefix('sdk:');
-      if (SdkTagValue.isNotAny(sdk)) {
-        tagsPredicate = tagsPredicate
-            .appendPredicate(TagsPredicate(requiredTags: ['sdk:$sdk']));
-      }
-    }
-    return FrontendSearchQuery._(
-      query: query ?? this.query,
-      tagsPredicate: tagsPredicate ?? this.tagsPredicate,
-      uploaderOrPublishers: uploaderOrPublishers ?? this.uploaderOrPublishers,
-      publisherId: publisherId ?? this.publisherId,
-      order: order ?? this.order,
-      offset: offset ?? this.offset,
-      limit: limit ?? this.limit,
-      includeDiscontinued: includeDiscontinued,
-      includeUnlisted: includeUnlisted,
-      includeLegacy: includeLegacy,
-    );
-  }
-
-  ServiceSearchQuery toServiceQuery() {
-    var tagsPredicate = this.tagsPredicate;
-    if (includeDiscontinued &&
-        tagsPredicate.isProhibitedTag(PackageTags.isDiscontinued)) {
-      tagsPredicate = tagsPredicate.withoutTag(PackageTags.isDiscontinued);
-    }
-    if (includeUnlisted &&
-        tagsPredicate.isProhibitedTag(PackageTags.isUnlisted)) {
-      tagsPredicate = tagsPredicate.withoutTag(PackageTags.isUnlisted);
-    }
-    if (includeLegacy &&
-        tagsPredicate.isProhibitedTag(PackageVersionTags.isLegacy)) {
-      tagsPredicate = tagsPredicate.withoutTag(PackageVersionTags.isLegacy);
-    }
-    return ServiceSearchQuery.parse(
-      query: query,
-      tagsPredicate: tagsPredicate,
-      uploaderOrPublishers: uploaderOrPublishers,
-      publisherId: publisherId,
-      offset: offset,
-      limit: limit,
-      order: order,
-    );
-  }
-
-  bool get hasQuery => query != null && query.isNotEmpty;
-
-  String get sdk {
-    final values = tagsPredicate._values.entries
-        .where((e) => e.key.startsWith('sdk:') && e.value == true)
-        .map((e) => e.key.split(':')[1]);
-    return values.isEmpty ? null : values.first;
-  }
-
-  /// Returns the validity status of the query.
-  QueryValidity evaluateValidity() {
-    // Block search on unreasonably long search queries (when the free-form
-    // text part is longer than one would enter via the search input field).
-    final queryLength = parsedQuery?.text?.length ?? 0;
-    if (queryLength > _maxQueryLength) {
-      return QueryValidity.reject(rejectReason: 'Query too long.');
-    }
-
-    // Consolidate the query to the state that the search service backend will
-    // recieve:
-    final backendQuery = toServiceQuery();
-
-    // Do not allow override of search filter tags. (E.g. search scope would
-    // require sdk:flutter, do not allow -sdk:flutter to override it).
-    final conflictingTags = backendQuery.tagsPredicate
-        ._getConflictingTags(backendQuery.parsedQuery.tagsPredicate);
-    if (conflictingTags.isNotEmpty) {
-      return QueryValidity.reject(
-          rejectReason:
-              'Tag conflict with search filters: `${conflictingTags.join(', ')}`.');
-    }
-    return QueryValidity.accept();
-  }
-
-  /// Converts the query to a user-facing link that the search form can use as
-  /// the base path of its `action` parameter.
-  String toSearchFormPath() {
-    String path = '/packages';
-    if (sdk != null) {
-      path = '/$sdk/packages';
-    }
-    if (tagsPredicate.isRequiredTag('is:flutter-favorite')) {
-      path = '/flutter/favorites';
-    }
-    if (publisherId != null && publisherId.isNotEmpty) {
-      path = '/publishers/$publisherId/packages';
-    }
-    if (uploaderOrPublishers != null && uploaderOrPublishers.isNotEmpty) {
-      path = myPackagesUrl();
-    }
-    return path;
-  }
-
-  /// Converts the query to a user-facing link that (after frontend parsing) will
-  /// re-create an identical search query object.
-  String toSearchLink({int page}) {
-    final params = <String, dynamic>{};
-    if (query != null && query.isNotEmpty) {
-      params['q'] = query;
-    }
-    params.addAll(tagsPredicate.asSearchLinkParams());
-    if (order != null) {
-      final String paramName = 'sort';
-      params[paramName] = serializeSearchOrder(order);
-    }
-    if (includeDiscontinued) {
-      params['discontinued'] = '1';
-    }
-    if (includeUnlisted) {
-      params['unlisted'] = '1';
-    }
-    if (includeLegacy && SdkTagValue.isAny(sdk)) {
-      params['legacy'] = '1';
-    }
-    if (page != null && page > 1) {
-      params['page'] = page.toString();
-    }
-    final path = toSearchFormPath();
-    if (params.isEmpty) {
-      return path;
-    } else {
-      return Uri(path: path, queryParameters: params).toString();
-    }
-  }
-}
-
 class ServiceSearchQuery {
   final String query;
-  final ParsedQuery parsedQuery;
+  final ParsedQueryText parsedQuery;
   final TagsPredicate tagsPredicate;
 
   /// The query will match packages where the owners of the package have
@@ -477,7 +244,7 @@ class ServiceSearchQuery {
     this.order,
     this.offset,
     this.limit,
-  })  : parsedQuery = ParsedQuery._parse(query),
+  })  : parsedQuery = ParsedQueryText.parse(query),
         tagsPredicate = tagsPredicate ?? TagsPredicate(),
         uploaderOrPublishers = _listToNull(uploaderOrPublishers),
         publisherId = _stringToNull(publisherId);
@@ -567,6 +334,28 @@ class ServiceSearchQuery {
         .where((e) => e.key.startsWith('sdk:') && e.value == true)
         .map((e) => e.key.split(':')[1]);
     return values.isEmpty ? null : values.first;
+  }
+
+  /// Returns the validity status of the query.
+  QueryValidity evaluateValidity() {
+    // Block search on unreasonably long search queries (when the free-form
+    // text part is longer than one would enter via the search input field).
+    final queryLength = parsedQuery?.text?.length ?? 0;
+    if (queryLength > _maxQueryLength) {
+      return QueryValidity.reject(rejectReason: 'Query too long.');
+    }
+
+    // Do not allow override of search filter tags. (E.g. search scope would
+    // require sdk:flutter, do not allow -sdk:flutter to override it).
+    final conflictingTags =
+        tagsPredicate._getConflictingTags(parsedQuery.tagsPredicate);
+    if (conflictingTags.isNotEmpty) {
+      return QueryValidity.reject(
+          rejectReason:
+              'Tag conflict with search filters: `${conflictingTags.join(', ')}`.');
+    }
+
+    return QueryValidity.accept();
   }
 }
 
@@ -735,7 +524,7 @@ class TagsPredicate {
   }
 }
 
-class ParsedQuery {
+class ParsedQueryText {
   final String text;
   final String packagePrefix;
 
@@ -754,7 +543,7 @@ class ParsedQuery {
   /// Detected tags in the user-provided query.
   TagsPredicate tagsPredicate;
 
-  ParsedQuery._(
+  ParsedQueryText._(
     this.text,
     this.packagePrefix,
     this.refDependencies,
@@ -764,7 +553,7 @@ class ParsedQuery {
     this.tagsPredicate,
   );
 
-  factory ParsedQuery._parse(String q) {
+  factory ParsedQueryText.parse(String q) {
     String queryText = q ?? '';
     queryText = ' $queryText ';
     String packagePrefix;
@@ -803,7 +592,7 @@ class ParsedQuery {
       queryText = null;
     }
 
-    return ParsedQuery._(
+    return ParsedQueryText._(
       queryText,
       packagePrefix,
       dependencies,
@@ -941,45 +730,4 @@ int extractPageFromUrlParameters(Map<String, String> queryParameters) {
   final pageAsString = queryParameters['page'];
   final pageAsInt = int.tryParse(pageAsString ?? '1') ?? 1;
   return max(pageAsInt, 1);
-}
-
-/// Parses the search query URL queryParameters for the parameters we expose on
-/// the frontend. The parameters and the values may be different from the ones
-/// we use in the search service backend.
-FrontendSearchQuery parseFrontendSearchQuery(
-  Map<String, String> queryParameters, {
-  String platform,
-  String sdk,
-  List<String> uploaderOrPublishers,
-  String publisherId,
-  @required TagsPredicate tagsPredicate,
-}) {
-  final int page = extractPageFromUrlParameters(queryParameters);
-  final int offset = resultsPerPage * (page - 1);
-  final String queryText = queryParameters['q'] ?? '';
-  final String sortParam = queryParameters['sort'];
-  final SearchOrder sortOrder = parseSearchOrder(sortParam);
-  List<String> runtimes;
-  if (queryParameters.containsKey('runtime')) {
-    runtimes = queryParameters['runtime'].split(' ');
-  }
-  List<String> platforms;
-  if (queryParameters.containsKey('platform')) {
-    platforms = queryParameters['platform'].split(' ');
-  }
-  return FrontendSearchQuery.parse(
-    query: queryText,
-    sdk: sdk,
-    runtimes: runtimes,
-    platforms: platforms,
-    uploaderOrPublishers: uploaderOrPublishers,
-    publisherId: publisherId,
-    order: sortOrder,
-    offset: offset,
-    limit: resultsPerPage,
-    includeDiscontinued: queryParameters['discontinued'] == '1',
-    includeUnlisted: queryParameters['unlisted'] == '1',
-    includeLegacy: queryParameters['legacy'] == '1' && SdkTagValue.isAny(sdk),
-    tagsPredicate: tagsPredicate,
-  );
 }
