+++ app/test/search/search_service_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:test/test.dart';
 
+import 'package:pub_dev/search/search_form.dart';
 import 'package:pub_dev/search/search_service.dart';
 
 void main() {
@@ -24,14 +25,13 @@ void main() {
 
   group('ParsedQuery', () {
     test('trim', () {
-      expect(FrontendSearchQuery.parse(query: 'text').parsedQuery.text, 'text');
-      expect(FrontendSearchQuery.parse(query: ' text ').query, 'text');
-      expect(
-          FrontendSearchQuery.parse(query: ' text ').parsedQuery.text, 'text');
+      expect(SearchForm.parse(query: 'text').parsedQuery.text, 'text');
+      expect(SearchForm.parse(query: ' text ').query, 'text');
+      expect(SearchForm.parse(query: ' text ').parsedQuery.text, 'text');
     });
 
     test('no dependency', () {
-      final query = FrontendSearchQuery.parse(query: 'text');
+      final query = SearchForm.parse(query: 'text');
       expect(query.parsedQuery.text, 'text');
       expect(query.parsedQuery.refDependencies, []);
       expect(query.parsedQuery.allDependencies, []);
@@ -39,7 +39,7 @@ void main() {
     });
 
     test('only one dependency', () {
-      final query = FrontendSearchQuery.parse(query: 'dependency:pkg');
+      final query = SearchForm.parse(query: 'dependency:pkg');
       expect(query.parsedQuery.text, isNull);
       expect(query.parsedQuery.refDependencies, ['pkg']);
       expect(query.parsedQuery.allDependencies, []);
@@ -47,7 +47,7 @@ void main() {
     });
 
     test('only one dependency*', () {
-      final query = FrontendSearchQuery.parse(query: 'dependency*:pkg');
+      final query = SearchForm.parse(query: 'dependency*:pkg');
       expect(query.parsedQuery.text, isNull);
       expect(query.parsedQuery.refDependencies, []);
       expect(query.parsedQuery.allDependencies, ['pkg']);
@@ -55,7 +55,7 @@ void main() {
     });
 
     test('two dependencies with text blocks', () {
-      final query = FrontendSearchQuery.parse(
+      final query = SearchForm.parse(
           query: 'text1 dependency:pkg1 text2 dependency:pkg2');
       expect(query.parsedQuery.text, 'text1 text2');
       expect(query.parsedQuery.refDependencies, ['pkg1', 'pkg2']);
@@ -64,7 +64,7 @@ void main() {
     });
 
     test('two mixed dependencies with text blocks', () {
-      final query = FrontendSearchQuery.parse(
+      final query = SearchForm.parse(
           query: 'text1 dependency:pkg1 text2 dependency*:pkg2');
       expect(query.parsedQuery.text, 'text1 text2');
       expect(query.parsedQuery.refDependencies, ['pkg1']);
@@ -73,34 +73,33 @@ void main() {
     });
 
     test('only publisher', () {
-      final query = FrontendSearchQuery.parse(query: 'publisher:example.com');
+      final query = SearchForm.parse(query: 'publisher:example.com');
       expect(query.parsedQuery.text, isNull);
       expect(query.parsedQuery.publisher, 'example.com');
     });
 
     test('only email', () {
-      final query = FrontendSearchQuery.parse(query: 'email:user@domain.com');
+      final query = SearchForm.parse(query: 'email:user@domain.com');
       expect(query.parsedQuery.text, isNull);
       expect(query.parsedQuery.emails, ['user@domain.com']);
     });
 
     test('known tag', () {
-      final query = FrontendSearchQuery.parse(query: 'is:legacy');
+      final query = SearchForm.parse(query: 'is:legacy');
       expect(query.parsedQuery.text, isNull);
       expect(
           query.parsedQuery.tagsPredicate.toQueryParameters(), ['is:legacy']);
     });
 
     test('forbidden known tag', () {
-      final query = FrontendSearchQuery.parse(query: '-is:legacy');
+      final query = SearchForm.parse(query: '-is:legacy');
       expect(query.parsedQuery.text, isNull);
       expect(
           query.parsedQuery.tagsPredicate.toQueryParameters(), ['-is:legacy']);
     });
 
     test('known tag + package prefix + search text', () {
-      final query =
-          FrontendSearchQuery.parse(query: 'json is:legacy package:foo_');
+      final query = SearchForm.parse(query: 'json is:legacy package:foo_');
       expect(query.parsedQuery.text, 'json');
       expect(
           query.parsedQuery.tagsPredicate.toQueryParameters(), ['is:legacy']);
@@ -108,7 +107,7 @@ void main() {
     });
 
     test('publisher + email + text + dependency', () {
-      final query = FrontendSearchQuery.parse(
+      final query = SearchForm.parse(
           query:
               'publisher:example.com email:user@domain.com text dependency:pkg1');
       expect(query.parsedQuery.text, 'text');
@@ -121,49 +120,48 @@ void main() {
 
   group('Search URLs', () {
     test('empty', () {
-      final query = FrontendSearchQuery.parse();
+      final query = SearchForm.parse();
       expect(query.parsedQuery.text, isNull);
       expect(query.parsedQuery.packagePrefix, isNull);
       expect(query.toSearchLink(), '/packages');
     });
 
     test('platform: flutter', () {
-      final query = FrontendSearchQuery.parse(sdk: 'flutter');
+      final query = SearchForm.parse(sdk: 'flutter');
       expect(query.parsedQuery.text, isNull);
       expect(query.parsedQuery.packagePrefix, isNull);
       expect(query.toSearchLink(), '/flutter/packages');
     });
 
     test('Flutter favorites', () {
-      final query = FrontendSearchQuery.parse(
+      final query = SearchForm.parse(
           tagsPredicate: TagsPredicate(requiredTags: ['is:flutter-favorite']));
       expect(query.toSearchLink(page: 2), '/flutter/favorites?page=2');
     });
 
     test('publisher: example.com', () {
-      final query = FrontendSearchQuery.parse(publisherId: 'example.com');
+      final query = SearchForm.parse(publisherId: 'example.com');
       expect(query.toSearchLink(), '/publishers/example.com/packages');
       expect(query.toSearchLink(page: 2),
           '/publishers/example.com/packages?page=2');
     });
 
     test('publisher: example.com with query', () {
-      final query =
-          FrontendSearchQuery.parse(publisherId: 'example.com', query: 'json');
+      final query = SearchForm.parse(publisherId: 'example.com', query: 'json');
       expect(query.toSearchLink(), '/publishers/example.com/packages?q=json');
       expect(query.toSearchLink(page: 2),
           '/publishers/example.com/packages?q=json&page=2');
     });
 
     test('package prefix: angular', () {
-      final query = FrontendSearchQuery.parse(query: 'package:angular');
+      final query = SearchForm.parse(query: 'package:angular');
       expect(query.parsedQuery.text, isNull);
       expect(query.parsedQuery.packagePrefix, 'angular');
       expect(query.toSearchLink(), '/packages?q=package%3Aangular');
     });
 
     test('complex search', () {
-      final query = FrontendSearchQuery.parse(
+      final query = SearchForm.parse(
           query: 'package:angular widget', order: SearchOrder.top);
       expect(query.parsedQuery.text, 'widget');
       expect(query.parsedQuery.packagePrefix, 'angular');
@@ -174,7 +172,7 @@ void main() {
 
   group('new sdk queries', () {
     test('sdk:flutter & platform:android', () {
-      final query = parseFrontendSearchQuery(
+      final query = parseFrontendSearchForm(
         {'platform': 'android'},
         sdk: 'flutter',
         tagsPredicate: TagsPredicate.regularSearch(),
@@ -193,7 +191,7 @@ void main() {
     });
 
     test('sdk:flutter & platform:android & platform:ios', () {
-      final query = parseFrontendSearchQuery(
+      final query = parseFrontendSearchForm(
         Uri.parse('/flutter/packages?platform=android++ios').queryParameters,
         sdk: 'flutter',
         tagsPredicate: TagsPredicate.regularSearch(),
@@ -213,7 +211,7 @@ void main() {
     });
 
     test('sdk:dart & runtime:web', () {
-      final query = parseFrontendSearchQuery(
+      final query = parseFrontendSearchForm(
         Uri.parse('/dart/packages?runtime=web').queryParameters,
         sdk: 'dart',
         tagsPredicate: TagsPredicate.regularSearch(),
